Fix bug with code coverage generation on Mac
This bug is caused by OffPiste/CMakeLists.txt, which checks generates coverage information using GCC flags (and only generates this information if GCC is the compiler). This led to a bug on Mac, where the compiler would appear to be clang, not GCC, so the coverage flags wouldn't get set.